[Refactor] refactoring for reducer
- Split one reducer into two auth and group reducers
