+++ src/containers/auth/LoginFormContainer.jsx
@@ -4,10 +4,10 @@ import { useUnmount } from 'react-use';
 import { useHistory } from 'react-router-dom';
 import { useSelector, useDispatch } from 'react-redux';
 
-import { get, isCheckValidate } from '../../util/utils';
+import { getAuth, isCheckValidate } from '../../util/utils';
 import {
   changeAuthField, clearAuth, clearAuthFields, requestLogin,
-} from '../../reducers/slice';
+} from '../../reducers/authSlice';
 import { ERROR_MESSAGE, FIREBASE_AUTH_ERROR_MESSAGE } from '../../util/messages';
 
 import AuthForm from '../../components/auth/AuthForm';
@@ -20,9 +20,9 @@ const LoginFormContainer = () => {
   const dispatch = useDispatch();
   const history = useHistory();
 
-  const login = useSelector(get('login'));
-  const user = useSelector(get('user'));
-  const authError = useSelector(get('authError'));
+  const login = useSelector(getAuth('login'));
+  const user = useSelector(getAuth('user'));
+  const authError = useSelector(getAuth('authError'));
 
   const onChangeLoginField = useCallback(({ name, value }) => {
     dispatch(
