+++ src/containers/common/HeaderContainer.jsx
@@ -3,8 +3,8 @@ import React, { useCallback } from 'react';
 import { useHistory } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { get } from '../../util/utils';
-import { requestLogout } from '../../reducers/slice';
+import { getAuth } from '../../util/utils';
+import { requestLogout } from '../../reducers/authSlice';
 
 import Header from '../../components/common/Header';
 
@@ -12,7 +12,7 @@ const HeaderContainer = () => {
   const dispatch = useDispatch();
   const history = useHistory();
 
-  const user = useSelector(get('user'));
+  const user = useSelector(getAuth('user'));
 
   const onLogout = useCallback(() => {
     dispatch(requestLogout());
