+++ src/pages/IntroducePage.jsx
@@ -3,6 +3,7 @@ import React from 'react';
 import { useParams } from 'react-router-dom';
 
 import Responsive from '../styles/Responsive';
+
 import IntroduceContainer from '../containers/introduce/IntroduceContainer';
 
 const IntroducePage = ({ params }) => {
