+++ src/reducers/rootSlice.js
@@ -1,0 +1,10 @@
+import { combineReducers } from '@reduxjs/toolkit';
+import authReducer from './authSlice';
+import groupReducer from './groupSlice';
+
+const rootReducer = combineReducers({
+  authReducer,
+  groupReducer,
+});
+
+export default rootReducer;
